Feat : 잔액 사용 취소 기능 구현 , 테스트 完

+++ src/main/java/com/ZB/demo/domain/Transaction.java
@@ -20,7 +20,7 @@ public class Transaction {
     private long id;
 
     private TransactionResult result;
-    private long transactionAmount;
+    private Integer transactionAmount;
     private String transactionAccountNumber;
     private LocalDateTime transactionDateTime;
